+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Change Log
 ### Bug Fixes
 * Aprimora fluxo de inserção de um modelo de checklist
 * Corrige busca dos modelos de checklist disponíveis
+* Lida com caso de alternativas nulas na inserção do modelo de checklist
 
 <a name="v0.0.57"></a>
 ## Version [v0.0.57](https://github.com/luizfp/PrologWeb/compare/v0.0.56...v0.0.57) (2018-05-08)